Clean arc master

+++ HumanitarianAssistance.Application/HR/Models/EmployeeSalaryAnalyticalInfoModel.cs
@@ -7,7 +7,7 @@ namespace HumanitarianAssistance.Application.HR.Models
    public class EmployeeSalaryAnalyticalInfoModel
     {
         public int EmployeeSalaryAnalyticalInfoId { get; set; }
-        public int? AccountCode { get; set; }
+        public long? AccountCode { get; set; }
         public long ProjectId { get; set; }
         public long BudgetLineId { get; set; }
         public double SalaryPercentage { get; set; }
